lavet ændringer i admin panelet

+++ admin/pages/delete-event.php
@@ -1,12 +1,21 @@
 <?php
 $id = $_GET['event_id'];
 $table = $_GET['table'];
-
-$sql = "DELETE FROM $table WHERE id = $id";
-$query = $handler->prepare($sql);
-if($query->execute()){
-	echo "der er slettet i databasen!";
-}else{
-	echo "Det virkede ikke!";
+if (isset($_GET['table'])) {
+	$sql = "DELETE FROM $table WHERE id = $id";
+	$query = $handler->prepare($sql);
+	$query->execute();
 }
-?>
+
+$sql = "DELETE FROM event WHERE id = $id";
+$query = $handler->prepare($sql);
+$query->execute();
+
+$sql = "DELETE FROM event_content WHERE event_id = $id";
+$query = $handler->prepare($sql);
+$query->execute();
+
+$sql = "DELETE FROM event_time WHERE event_id = $id";
+$query = $handler->prepare($sql);
+$query->execute();
+header('Location: ' . $_SERVER['HTTP_REFERER']);
